deepfake 1st project/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch==2.1.1
torchvision==0.16.1
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.0.0
sqlalchemy==2.0.23
sqlmodel==0.0.14
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
cors==1.0.1
gradio==4.40.0
